# Update file paths in tables only once