Closes #145

This uses a single place to update the file paths and the new `audformat` util functions.
The advantage is that now all updates are included in the progress bar,
and we have a single place in `audformat` to maintain the functions.
The execution time increases slightly as we add a little bit more overhead in using the `audformat` util functions, but 0.1 s is very marginal.

I did the following benchmarks on my local PC.

| branch | audformat | time (average over 20  runs) |
| --------- | -------------- | ------------------------------------- |
| master | 0.12.1 | 5.87 s |
| master | 0.12.2 | 5.82 s |
| this | 0.12.1 | 19.27 s |
| this | 0.12.2 | 5.96 s |