@frankenjoe 2021-11-18
> On `timedeltas.lambda()` we could improve as outlined in #144, but at the moment I'm not too motivated to tackle this before the next release.

Yes, I am also not sure if it's really worth the trouble for an additional 1.6 s we safe for an already quite large database like `librispeech`.